[dynamo] Raise exception on incorrect usage of disallow_in_graph
[ghstack-poisoned]
