+++ test/dynamo/test_misc.py
@@ -23,7 +23,7 @@ import torch._dynamo.test_case
 import torch._dynamo.testing
 import torch.onnx.operators
 from torch._C import FileCheck
-from torch._dynamo import bytecode_transformation, graph_break
+from torch._dynamo import bytecode_transformation
 from torch._dynamo.output_graph import OutputGraph
 from torch._dynamo.testing import (
     CompileCounter,
@@ -48,10 +48,6 @@ from torch.testing._internal.common_utils import freeze_rng_state
 from torch.testing._internal.jit_utils import JitTestCase
 
 mytuple = collections.namedtuple("mytuple", ["a", "b", "ab"])
-
-
-def my_custom_function(x):
-    return x + 1
 
 
 class MyPickledModule(torch.nn.Module):
@@ -1697,25 +1693,6 @@ class MiscTests(torch._dynamo.test_case.TestCase):
         torch._dynamo.run()(fn2)(torch.randn(4))
         self.assertEqual(cnts2.frame_count, 0)
 
-    def test_graph_break(self):
-        cnts = torch._dynamo.testing.CompileCounter()
-
-        @torch._dynamo.optimize(cnts)
-        def fn(x):
-            x = torch.cos(x)
-            x = torch.cos(x)
-            torch._dynamo.graph_break()
-            x = torch.cos(x)
-            x = torch.cos(x)
-            graph_break()
-            x = torch.cos(x)
-            x = torch.cos(x)
-            return x
-
-        fn(torch.randn(4, 5))
-        self.assertEqual(cnts.frame_count, 3)
-        self.assertEqual(cnts.op_count, 6)
-
     def test_torch_size(self):
         cnts = torch._dynamo.testing.CompileCounter()
 
@@ -2355,46 +2332,6 @@ def fn():
         opt_f4 = torch._dynamo.optimize(cnts)(f4)
         res2 = opt_f4()
         self.assertTrue(same(res1, res2))
-
-    def test_disallow_in_graph(self):
-        cnts = torch._dynamo.testing.CompileCounter()
-
-        @torch._dynamo.optimize(cnts)
-        def fn(a):
-            x = torch.add(a, 1)
-            x = torch.add(x, 1)
-            x = torch.sub(x, 1)
-            x = torch.add(x, 1)
-            x = torch.add(x, 1)
-            return x
-
-        torch._dynamo.disallow_in_graph(torch.sub)
-        fn(torch.randn(10))
-        torch._dynamo.allow_in_graph(torch.sub)
-
-        # check for graph break on sub
-        self.assertEqual(cnts.frame_count, 2)
-        self.assertEqual(cnts.op_count, 4)
-
-    def test_allow_in_graph(self):
-        cnts = torch._dynamo.testing.CompileCounter()
-
-        @torch._dynamo.optimize(cnts)
-        def fn(a):
-            x = torch.add(a, 1)
-            x = torch.add(x, 1)
-            x = my_custom_function(x)
-            x = torch.add(x, 1)
-            x = torch.add(x, 1)
-            return x
-
-        torch._dynamo.allow_in_graph(my_custom_function)
-        fn(torch.randn(10))
-        torch._dynamo.disallow_in_graph(my_custom_function)
-
-        # check for no graph break
-        self.assertEqual(cnts.frame_count, 1)
-        self.assertEqual(cnts.op_count, 5)
 
     def test_sample_input(self):
         from torch.testing._internal.common_methods_invocations import SampleInput
