+++ torch/_dynamo/exc.py
@@ -97,6 +97,10 @@ class UserError(Unsupported):
         self.error_type = error_type
 
 
+class IncorrectUsage(Exception):
+    pass
+
+
 def unimplemented(msg: str):
     assert msg != os.environ.get("BREAK", False)
     raise Unsupported(msg)
